WIP move send_email functionality to data_request_area
