+++ app/views/cases/commissioning_documents/new.html.slim
@@ -24,7 +24,14 @@ span.visually-hidden
 
     - url = case_data_request_area_data_request_commissioning_documents_path(@case, @data_request_area, @data_request)
     = form_for @commissioning_document, url: url do |f|
-      = render 'form', f: f
+      
+      = f.hidden_field :template_name, value: ""
+      = f.radio_button_fieldset :template_name, choices: CommissioningDocument::TEMPLATE_TYPES.keys
+      - if FeatureSet.email_commissioning_document.enabled?
+        = f.submit t('button.generate_document'), class: 'button', data: { disable_with: false }
+      - else
+        = f.submit t('button.download_document'), class: 'button', data: { disable_with: false }
+
