+++ app/views/cases/data_request_areas/send_email.html.slim
@@ -39,7 +39,7 @@ span.visually-hidden
       div
         = raw @recipient_emails.join("<br>").html_safe
 
-    div.button-holder
+    / div.button-holder
       = form_for @commissioning_document, method: :post, url: send_email_case_data_request_area_data_request_commissioning_document_path(@case, @data_request_area, @data_request, @commissioning_document) do |f|
         = submit_tag t('button.send_email'), class: 'button data_request_send_email', disabled: @no_email_present
-        = link_to(t('common.cancel'), redirect_to case_data_request_area_path(@case, @data_request_area), class: 'acts-like-button data_request_cancel moj-style-link')
+        = link_to(t('common.cancel'), case_data_request_area_path(@case, @data_request_area), class: 'acts-like-button data_request_cancel moj-style-link')
