+++ app/views/cases/data_request_areas/show.html.slim
@@ -27,3 +27,40 @@ div class="case"
 
   section.case-info
     = render partial: 'cases/offender_sar/data_requests', locals: { case_details: @case, data_request_area: @data_request_area, data_request: @data_request, allow_editing: allow_editing }
+
+    div.commissioning-document
+      h2.request--heading
+        = t('cases.data_request_areas.show.emails_header')
+      table.report
+        thead
+          tr
+            th = t('cases.commissioning_documents.table_headings.request_document')
+            th = t('cases.commissioning_documents.table_headings.last_updated')
+            th = t('cases.commissioning_documents.table_headings.action')
+        tbody.document
+          tr
+            td = @data_request_area.request_document
+            td = @commissioning_document&.updated_at
+            td
+              span = @commissioning_document&.download_link
+      - unless @commissioning_document&.sent? || !policy(@case).can_record_data_request?
+        div.button-holder
+        = link_to t('button.send_email'), send_email_case_data_request_area_path(@case, @data_request_area), class: 'button button-high data_request_send_email'
+
+      - if @data_request_area.data_request_emails.present?
+        h2.heading-medium
+          = t('.emails_history_header')
+        table.report
+          thead
+            tr
+              th = t('cases.commissioning_documents.email_history_table_headings.email_type')
+              th = t('cases.commissioning_documents.email_history_table_headings.email_address')
+              th = t('cases.commissioning_documents.email_history_table_headings.date_sent')
+              th = t('cases.commissioning_documents.email_history_table_headings.status')
+          tbody.email_details
+            - @data_request_area.data_request_emails.map(&:decorate).each do |email|
+              tr
+                td = email.email_type
+                td = email.email_address
+                td = email.created_at
+                td = email.status
