+++ spec/controllers/cases/data_request_areas_controller_spec.rb
@@ -107,4 +107,123 @@ RSpec.describe Cases::DataRequestAreasController, type: :controller do
       end
     end
   end
+
+  xdescribe "#send_email" do
+    let(:data_request) do
+      create(
+        :data_request,
+        cached_num_pages: 10,
+        completed: true,
+        cached_date_received: Time.zone.yesterday,
+        commissioning_document:,
+        )
+    end
+    let(:params) do
+      {
+        id: data_request.id,
+        case_id: data_request.case_id,
+        data_request_area_id: data_request_area.id,
+      }
+    end
+    let(:commissioning_document) { create(:commissioning_document, template_name:) }
+    let(:template_name) { "prison" }
+
+    it "assigns value to recipient emails" do
+      allow_any_instance_of(DataRequest) # rubocop:disable RSpec/AnyInstance
+        .to receive(:recipient_emails).and_return("test@email.com")
+      get(:send_email, params:)
+      expect(assigns(:recipient_emails)).to eq("test@email.com")
+    end
+
+    context "with no associated email" do
+      it "returns no associated email present" do
+        allow_any_instance_of(DataRequest) # rubocop:disable RSpec/AnyInstance
+          .to receive(:recipient_emails).and_return([])
+        get(:send_email, params:)
+        expect(assigns(:no_email_present)).to eq(true)
+      end
+    end
+
+    context "when probation document selected" do
+      let(:template_name) { "probation" }
+
+      it "routes to the send_email branston probation page" do
+        get(:send_email, params:)
+        expect(response).to render_template(:probation_send_email)
+      end
+
+      context "with confirm probation email" do
+        let(:params) do
+          {
+            id: data_request.id,
+            case_id: data_request.case_id,
+            data_request_area_id: data_request_area.id,
+            probation_commissioning_document_email: {
+              probation: 1,
+              email_branston_archives: "yes",
+            },
+          }
+        end
+
+        it "adds the branston probation email to recipients" do
+          post(:send_email, params:)
+          expect(response).to render_template(:send_email)
+          expect(assigns(:recipient_emails)).to include(CommissioningDocumentTemplate::Probation::BRANSTON_ARCHIVES_EMAIL)
+        end
+
+        it "updates the data_request" do
+          post(:send_email, params:)
+          expect(data_request.reload.email_branston_archives).to be_truthy
+        end
+      end
+
+      context "with decline probation email" do
+        let(:params) do
+          {
+            id: data_request.id,
+            case_id: data_request.case_id,
+            data_request_area_id: data_request_area.id,
+            probation_commissioning_document_email: {
+              probation: 1,
+              email_branston_archives: "no",
+            },
+          }
+        end
+
+        it "doesnt add the branston probation email to recipients" do
+          post(:send_email, params:)
+          expect(response).to render_template(:send_email)
+          expect(assigns(:recipient_emails)).not_to include(CommissioningDocumentTemplate::Probation::BRANSTON_ARCHIVES_EMAIL)
+        end
+      end
+
+      context "with no options chosen" do
+        let(:params) do
+          {
+            id: data_request.id,
+            case_id: data_request.case_id,
+            data_request_area_id: data_request_area.id,
+            probation_commissioning_document_email: {
+              probation: 1,
+            },
+          }
+        end
+
+        it "raises error message" do
+          post(:send_email, params:)
+          expect(response).to render_template(:probation_send_email)
+          expect(assigns(:email)).not_to be_valid
+        end
+      end
+    end
+
+    context "with non-probation document" do
+      let(:template_name) { "prison" }
+
+      it "routes to the send_email confirmation page" do
+        get(:send_email, params:)
+        expect(response).to render_template(:send_email)
+      end
+    end
+  end
 end
